Add anisotropic thermal conduction, disabling fluid evol, fix/extend div B cleaning.
Key changes:

1. Anisotropic thermal conduction
Add supports for both a fixed thermal diffusivity coefficient as well as
Spitzer based thermal conductivty. See `doc/input.md` for parameters and usage.
Also added two test cases: a convergence test for ring diffusion in 2D
and a multi-dimensional test for diffusion of a hot blob in all planes.

2. Disabling fluid evol
If desired (e.g., for testing diffusive processes alone) the hyperbolic fluxes
can now be disabled by choosing the `"none"` Riemann solver in the input file.
Effectively, this calls a Riemann solver that sets all fluxes to 0 so that other
physics can still modify the flux arrays (but starting with a clean slate).

3. Fix/extend DivB cleaning
Given the prim to cons conversion of the psi scalar and it's initial value of 0
the original version did not clean the divergence as expected.
Now the psi array is properly updated.
In addition, the originally fixed parameter corresponding to the ratio of
transport to diffusion of DivB (as given by Dedner) has been updated to
use a free parameter (following Mignone & Tzeferacos 2010), which also
takes into account grid spacing, see discussion in `doc/input.md`.

Minor changes:
- Added a field loop convergence test.
- Started to populated `doc/input.md` with information specific to the
implementation choices in `AthenaPK`